Introduce ClientRepository interface and InMemoryClientRepository

+++ src/main/java/main/ClientService.java
@@ -1,13 +1,19 @@
 package main;
 
-import java.util.ArrayList;
 import java.util.List;
 public class ClientService {
 
-    private ClientRepository repository = new ClientRepository();
+    private final ClientRepository repository;
     private int nextId = 1;
 
-    private void ValidateClientData(String name, String phone, String email){
+    public ClientService(ClientRepository repository){
+        this.repository = repository;
+        this.nextId = repository.findAll().stream()
+                .mapToInt(Client::getId)
+                .max()
+                .orElse(0)+1;
+    }
+    private void validateClientData(String name, String phone, String email){
         if(name == null || name.isBlank()) {
             throw new IllegalArgumentException("Name can not be empty");
         }
@@ -21,22 +27,22 @@ public class ClientService {
         }
     }
 
-    private void CheckUniqueNumber(String phone){
+    private void checkUniqueNumber(String phone){
         if(repository.findByPhone(phone) != null){
             throw new IllegalArgumentException("Client with that number already exists");
         }
     }
 
     public void addClient(String name, String phone, String email){
-        ValidateClientData(name, phone, email);
-        CheckUniqueNumber(phone);
+        validateClientData(name, phone, email);
+        checkUniqueNumber(phone);
         Client client = new Client.ClientBuilder()
                 .SId(nextId)
                 .SName(name)
                 .SPhone(phone)
                 .SEmail(email)
                 .build();
-        repository.Save(client);
+        repository.save(client);
         nextId++;
     }
 
@@ -49,21 +55,22 @@ public class ClientService {
     }
 
     public void delete(Client client){
-        repository.delete(client);
+        if(client != null){ repository.delete(client); }
+        else return;
     }
 
-    public boolean updateClient(int id, Client updatedClient){
-        ValidateClientData(
-                updatedClient.getName(),
-                updatedClient.getPhone(),
-                updatedClient.getEmail()
+    public boolean updateClient(int id, String name, String phone, String email){
+        validateClientData(
+                name,
+                phone,
+                email
         );
 
         Client client = findById(id);
         if(client == null) return false;
-        client.setName(updatedClient.getName());
-        client.setPhone(updatedClient.getPhone());
-        client.setEmail(updatedClient.getEmail());
+        client.setName(name);
+        client.setPhone(phone);
+        client.setEmail(email);
         return true;
     }
